Move e2e for egressIP into new file under test/e2e
This commit moves the egressIP e2e's to a new file to
avoid cluttering the main e2e.go which is already a
mess in regards to functionality wise split of tests.
